feat: some more graphical logic and structures

+++ include/kernel/terminal.hxx
@@ -20,7 +20,7 @@ public:
 	size_t row;
 	size_t column;
     static Terminal &instance();
-    void put_entry_at(char c, uint8_t color, size_t x, size_t y);
+    void put_entry_at(char c, uint32_t color, size_t x, size_t y);
     void put_char(char c, uint8_t color);
     void write(const char* data, size_t size);
     void write(const char* data);
